finalização dos testes unitarios de banco de dados

+++ pkg/entity/Log.go
@@ -5,12 +5,12 @@ import "time"
 type Log struct {
 	Id        int64
 	UsuarioId int64
-	EventoId  int
+	EventoId  int64
 	Message   string
 	Data      time.Time
 }
 
-func NewLog(usuarioId int64, eventoId int, message string) *Log {
+func NewLog(usuarioId int64, eventoId int64, message string) *Log {
 	l := new(Log)
 	l.Message = message
 	l.UsuarioId = usuarioId
